split IconFetcher
